Update services.xml
Without this change an export of an emotion world is not possible. 

core.ERROR: Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException: The "translation" service or alias has been removed or inlined when the container was compiled. You should either make it public, or stop using the container directly and use dependency injection instead.